backend/requirements.txt
fastapi
uvicorn
python-dotenv
autogen==0.3.0
langchain_community
langchain_openai